requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
pymongo==4.7.1
aiohttp==3.9.5
python-dotenv==1.0.1
bip-utils==2.9.2
